Fix(split_income): Convert positions to weight before subtracting taxes (#8)
Previously, the code assumed that income and taxes used the same
currency. If the postings had an additional associated cost or price,
this was not considered. Now, to obtain the correct amount, the
beancount.core.convert.get_weight function is used. It provides the
amount that will need to be balanced from a posting of a transaction.

Appropriate tests are included.

Fixes #7